agent-AI/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
mistralai==0.3.0
google-generativeai==0.5.4